Fix the cache config to handle zstd compressed packages.
Update the location regex for the cache block to match requests for
packages with the .tar.xz extension or the .tar.zst extension.

Pacman has added support for zstd compression and the first zstd
compressed packages have started to appear on the package mirrors.

Those packages use the .tar.zst extension. The location regex matched
only requests for packages with the .tar.xz extension and as a result
the cache server returned a 404 error for zstd compressed packages.